gemma2-2b: fp32->fp16 (fp32 OOMs T4 VRAM); coherence-gated for fp16 soft-cap NaN
gemma2-2b loaded fine after the RAM fix but fp32 (~10GB + activations) exceeds the
T4's ~14.5GB VRAM. fp16 is the only fit on T4 (no bf16). Gemma-2 soft-capping may be
unstable in fp16 — quick_coherence_check gates it: if output is garbage, drop the
model (Gemma sweep falls back to G1 + G3).

+++ src/safety_circuits/config.py
@@ -101,8 +101,11 @@ MODELS: dict[str, ModelSpec] = {
         key="gemma2-2b",
         hf_name="google/gemma-2-2b-it",
         tl_name=None,  # GATED: accept terms at hf.co/google/gemma-2-2b-it
+        # fp32 (~10GB) + activations OOMs the T4's ~14.5GB VRAM, so fp16 is forced.
+        # Gemma-2's logit soft-capping *can* be unstable in fp16 — the coherence check
+        # gates this: if completions are garbage/NaN, drop gemma2-2b (keep G1 + G3).
         refusal_first_tokens=_REFUSAL,
-        dtype="float32",  # Gemma-2 soft-capping is unstable in fp16; T4 has no bf16
+        dtype="float16",
     ),
     "gemma3-1b": ModelSpec(
         key="gemma3-1b",
